@@ -1,6 +1,6 @@
 {
   "name": "sandbox",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "private": false,
   "engines": {
     "node": ">=16.0.0",
@@ -16,59 +16,59 @@
     "patch:hosts": "fec patch-etc-hosts",
     "start": "HOT=true fec dev",
     "test": "jest",
-    "postinstall": "rimraf .cache",
+    "postinstall": "ts-patch install && rimraf .cache",
     "verify": "npm-run-all build lint test"
   },
   "dependencies": {
-    "@patternfly/react-core": "4.250.1",
-    "@patternfly/react-table": "4.111.33",
-    "@redhat-cloud-services/frontend-components": "^3.9.25",
-    "@redhat-cloud-services/frontend-components-notifications": "^3.2.12",
-    "@redhat-cloud-services/frontend-components-utilities": "^3.3.7",
+    "@patternfly/react-component-groups": "^1.0.2",
+    "@patternfly/react-core": "^5.0.0",
+    "@patternfly/react-icons": "^5.0.0",
+    "@redhat-cloud-services/frontend-components": "^4.0.7",
+    "@redhat-cloud-services/frontend-components-notifications": "4.0.0",
+    "@redhat-cloud-services/frontend-components-utilities": "^4.0.0",
     "axios": "^0.27.2",
-    "classnames": "^2.3.1",
+    "classnames": "^2.3.2",
     "lodash-es": "^4.17.21",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-redux": "7.2.9",
-    "react-router-dom": "^6.12.0",
-    "redux": "4.2.0",
-    "redux-logger": "3.0.6",
-    "redux-promise-middleware": "6.1.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.1.2",
+    "react-router-dom": "^6.15.0",
+    "redux": "^4.2.1",
+    "redux-logger": "^3.0.6",
+    "redux-promise-middleware": "^6.1.3"
   },
   "devDependencies": {
-    "@babel/core": "7.14.0",
-    "@babel/plugin-proposal-object-rest-spread": "7.14.7",
-    "@babel/plugin-transform-runtime": "7.14.5",
-    "@babel/preset-env": "7.14.7",
-    "@babel/preset-react": "7.14.5",
-    "@babel/preset-typescript": "^7.17.12",
+    "@babel/core": "^7.22.15",
+    "@babel/preset-env": "^7.22.15",
+    "@babel/preset-react": "^7.22.15",
+    "@babel/preset-typescript": "^7.22.15",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.8",
-    "@redhat-cloud-services/eslint-config-redhat-cloud-services": "^1.2.6",
-    "@redhat-cloud-services/frontend-components-config": "^4.6.33",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/react-hooks": "^8.0.1",
+    "@redhat-cloud-services/eslint-config-redhat-cloud-services": "^2.0.2",
+    "@redhat-cloud-services/frontend-components-config": "^6.0.2",
+    "@redhat-cloud-services/tsc-transform-imports": "^1.0.4",
+    "@testing-library/jest-dom": "^6.1.3",
+    "@testing-library/react": "^14.0.0",
+    "@types/jest": "^29.5.4",
     "@types/lodash-es": "^4.17.7",
-    "@types/react": "17.0.39",
+    "@types/react": "^18.2.0",
     "@types/react-dom": "^18.0.5",
     "@types/react-router-dom": "^5.3.3",
     "@types/redux-logger": "^3.0.9",
     "@typescript-eslint/eslint-plugin": "^5.29.0",
     "@typescript-eslint/parser": "^5.29.0",
-    "babel-jest": "27.0.5",
-    "babel-plugin-transform-imports": "^2.0.0",
     "eslint": "8.18.0",
     "identity-obj-proxy": "3.0.0",
-    "jest": "27.0.5",
+    "jest": "29.6.4",
+    "jest-environment-jsdom": "^29.6.4",
     "npm-run-all": "4.1.5",
     "prop-types": "15.7.2",
     "rimraf": "^3.0.2",
     "stylelint": "13.13.1",
     "stylelint-config-recommended-scss": "4.2.0",
     "stylelint-scss": "3.19.0",
+    "ts-patch": "^3.0.2",
     "typesafe-actions": "^5.1.0",
-    "typescript": "^4.7.4",
+    "typescript": "^5.1.6",
     "webpack-bundle-analyzer": "4.4.2"
   },
   "insights": {
