@@ -1,7 +1,11 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import AppEntry from './AppEntry';
 
-const root = document.getElementById('root');
+const container = document.getElementById('root');
 
-ReactDOM.render(<AppEntry />, root, () => root?.setAttribute('data-ouia-safe', 'true'));
+if (container) {
+  container.setAttribute('data-ouia-safe', 'true');
+  const root = createRoot(container);
+  root.render(<AppEntry />);
+}
