@@ -11,15 +11,14 @@ module.exports = {
    * Add additional webpack plugins
    */
   plugins: [],
-  _unstableHotReload: process.env.HOT === 'true',
+  hotReload: process.env.HOT === 'true',
   moduleFederation: {
     exclude: ['react-router-dom'],
     shared: [
       {
         'react-router-dom': {
           singleton: true,
-          import: false,
-          requiredVersion: '^6.3.0',
+          version: '^6.8.0',
         },
       },
     ],
