@@ -1,14 +1,12 @@
 import * as React from 'react';
-import {
-  Divider,
-  Flex,
-  FlexItem,
-  PageSection,
-  PageSectionVariants,
-  Text,
-  TextContent,
-  TextVariants,
-} from '@patternfly/react-core';
+import { Divider } from '@patternfly/react-core/dist/dynamic/components/Divider';
+import { Flex } from '@patternfly/react-core/dist/dynamic/layouts/Flex';
+import { FlexItem } from '@patternfly/react-core/dist/dynamic/layouts/Flex';
+import { PageSection } from '@patternfly/react-core/dist/dynamic/components/Page';
+import { PageSectionVariants } from '@patternfly/react-core/dist/dynamic/components/Page';
+import { Text } from '@patternfly/react-core/dist/dynamic/components/Text';
+import { TextContent } from '@patternfly/react-core/dist/dynamic/components/Text';
+import { TextVariants } from '@patternfly/react-core/dist/dynamic/components/Text';
 
 type Props = {
   children: React.ReactNode;
