@@ -1,5 +1,6 @@
 import * as React from 'react';
-import { Gallery, GalleryItem } from '@patternfly/react-core';
+import { Gallery } from '@patternfly/react-core/dist/dynamic/layouts/Gallery';
+import { GalleryItem } from '@patternfly/react-core/dist/dynamic/layouts/Gallery';
 import { useSandboxServices } from '../../hooks/useSandboxServices';
 import ServiceCard from './ServiceCard';
 
