@@ -1,8 +1,8 @@
 import React, { useEffect } from 'react';
 
 import { Main } from '@redhat-cloud-services/frontend-components/Main';
-import { NotAuthorized } from '@redhat-cloud-services/frontend-components/NotAuthorized';
 import { useChrome } from '@redhat-cloud-services/frontend-components/useChrome';
+import { NotAuthorized } from '@patternfly/react-component-groups';
 
 const NoPermissionsPage = () => {
   const { appAction } = useChrome();
